shimono-tools/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=5.0.0
fonttools>=4.40.0
brotli>=1.0.9
